#27-#33: refact transferred type
