+++ internal/http/handler/site_dto.go
@@ -15,21 +15,3 @@ type createSiteRequest struct {
 	URL  string `json:"url"`
 	Name string `json:"name,omitempty"`
 }
-
-type SiteCheckHistoryItem struct {
-    ID           string    `json:"id"`
-    IsAvailable  bool      `json:"is_available"`
-    HTTPStatus   *int      `json:"http_status,omitempty"`
-    ResponseTime *int      `json:"response_time_ms,omitempty"`
-    ErrorMessage *string   `json:"error_message,omitempty"`
-    CheckedAt    time.Time `json:"checked_at"`
-}
-
-
-// PaginatedResponse — универсальный ответ с пагинацией
-type PaginatedResponse[T any] struct {
-    Data  []T `json:"data"`
-    Total int `json:"total"`
-    Limit int `json:"limit"`
-    Offset int `json:"offset"`
-}
